moved nextOnRequestError function and tests into next folder
